feat(LightingScenario): Add support for baking lightmap scenarios
WHY: Unity does not support baking and switching different lightmaps.

By using Unitys Lightmapping API, we bake the scene after we
activate a scenario (set specified Renderer and Light objects active).
 Then we move the baked files into a scenario directory, save lightmap
info per object, update references to lightmap assets.
 Activating a scenario uses these assets and records to reapply the
lightmap. Note: Lighting manager must not be in the same prefab as the renderers,
as the references will then be prefab references and we need
scene references for Unity to get what is going on.